% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{asegurar_orientacion_antihoraria}
\alias{asegurar_orientacion_antihoraria}
\title{Asegura que todas las geometrias de un objeto sf tengan orientacion antihoraria (CCW)}
\usage{
asegurar_orientacion_antihoraria(sf_obj)
}
\arguments{
\item{sf_obj}{Objeto sf.}
}
\value{
Objeto sf con orientaciones corregidas.
}
\description{
Asegura que todas las geometrias de un objeto sf tengan orientacion antihoraria (CCW)
}
