% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{corregir_poligono_ccw}
\alias{corregir_poligono_ccw}
\title{Corrige la orientacion de un poligono sf para que el anillo exterior sea CCW
y los anillos interiores (huecos) sean CW.}
\usage{
corregir_poligono_ccw(poly)
}
\arguments{
\item{poly}{Objeto de tipo polygon de sf (lista de matrices).}
}
\value{
Objeto de tipo polygon corregido.
}
\description{
Corrige la orientacion de un poligono sf para que el anillo exterior sea CCW
y los anillos interiores (huecos) sean CW.
}
