% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peruocc-package.R
\docType{package}
\name{peruocc-package}
\alias{peruocc}
\alias{peruocc-package}
\title{peruocc: Query and Standardize Biodiversity Occurrences in Peru}
\description{
Facilitates the retrieval, spatial validation, and integration of flora and fauna occurrence records across administrative units (districts and provinces) in Peru. Retrieves official boundary geometries via 'geoperu', queries and consolidates observations from the Global Biodiversity Information Facility (GBIF, \url{https://www.gbif.org/}) and 'iNaturalist' (\url{https://www.inaturalist.org/}), and standardizes attributes into a unified Darwin Core aligned structure. Designed for biodiversity assessments and spatial workflows within user-defined areas of interest.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paulesantos.github.io/peruocc/}
  \item Report bugs at \url{https://github.com/PaulESantos/peruocc/issues}
}

}
\author{
\strong{Maintainer}: Paul E. Santos Andrade \email{paulefrens@gmail.com} (\href{https://orcid.org/0000-0002-6635-0375}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Paul E. Santos Andrade \email{paulefrens@gmail.com} (\href{https://orcid.org/0000-0002-6635-0375}{ORCID}) [copyright holder]
}

}
\keyword{internal}
