% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{poligono_a_wkt}
\alias{poligono_a_wkt}
\title{Convierte un objeto sf a formato WKT (Well-Known Text) para consultas GBIF}
\usage{
poligono_a_wkt(sf_obj)
}
\arguments{
\item{sf_obj}{Objeto sf.}
}
\value{
Una cadena de texto en formato WKT.
}
\description{
Convierte un objeto sf a formato WKT (Well-Known Text) para consultas GBIF
}
