% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{simplificar_para_api}
\alias{simplificar_para_api}
\title{Simplifica un poligono sf o genera su bounding box si es muy complejo
para cumplir con el limite de longitud de caracteres de WKT.}
\usage{
simplificar_para_api(sf_obj, max_char = 1500, tolerancia_inicial_metros = 100)
}
\arguments{
\item{sf_obj}{Objeto sf.}

\item{max_char}{Limite de caracteres WKT (def: 1500).}

\item{tolerancia_inicial_metros}{Tolerancia inicial en metros para la simplificacion.}
}
\value{
Objeto sf simplificado (o su bbox) apto para consulta.
}
\description{
Simplifica un poligono sf o genera su bounding box si es muy complejo
para cumplir con el limite de longitud de caracteres de WKT.
}
