% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{simplificar_poligono}
\alias{simplificar_poligono}
\title{Simplifica un poligono de tipo sf usando proyecciones UTM}
\usage{
simplificar_poligono(sf_obj, tolerancia_metros = 100)
}
\arguments{
\item{sf_obj}{Objeto sf con la geometria.}

\item{tolerancia_metros}{Tolerancia de simplificacion en metros (def: 100 metros).}
}
\value{
Objeto sf simplificado en EPSG:4326.
}
\description{
Simplifica un poligono de tipo sf usando proyecciones UTM
}
