% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_combined.R
\name{buscar_especies_distrito}
\alias{buscar_especies_distrito}
\title{Realiza una busqueda combinada de especies en un distrito usando GBIF e iNaturalist}
\usage{
buscar_especies_distrito(
  distrito,
  departamento = NULL,
  provincia = NULL,
  nombre_cientifico = NULL,
  grupo = NULL,
  limite_por_api = configuracion_predeterminada()$limite_por_api,
  guardar_resultados = FALSE,
  tolerancia_simplificacion = configuracion_predeterminada()$tolerancia_simplificacion_m
)
}
\arguments{
\item{distrito}{Nombre del distrito (requerido).}

\item{departamento}{Nombre del departamento (opcional).}

\item{provincia}{Nombre de la provincia (opcional).}

\item{nombre_cientifico}{Nombre de la especie o grupo taxonomico a filtrar (opcional).}

\item{grupo}{Grupo taxonomico a filtrar: "flora", "fauna" o NULL.}

\item{limite_por_api}{Numero maximo de registros a solicitar por API (def: 500).}

\item{guardar_resultados}{Logico; si es TRUE guarda los resultados en processed/ (def: FALSE).}

\item{tolerancia_simplificacion}{Tolerancia en metros para simplificar el poligono en GBIF (def: 100).}
}
\value{
Una lista con el poligono del distrito, el dataframe de ocurrencias y estadisticas de resumen.
}
\description{
Realiza una busqueda combinada de especies en un distrito usando GBIF e iNaturalist
}
