% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_gbif.R
\name{buscar_gbif_por_poligono}
\alias{buscar_gbif_por_poligono}
\title{Busca ocurrencias de especies en GBIF dentro de un poligono de distrito}
\usage{
buscar_gbif_por_poligono(
  poligono_sf,
  nombre_cientifico = NULL,
  grupo = NULL,
  limite = 500,
  tolerancia_simplificacion = 100
)
}
\arguments{
\item{poligono_sf}{Objeto sf que representa el distrito (EPSG:4326).}

\item{nombre_cientifico}{Nombre de la especie o grupo taxonomico (opcional).}

\item{grupo}{Grupo taxonomico: "flora", "fauna" o NULL.}

\item{limite}{Numero maximo de registros a recuperar (max. 100000 para occ_search).}

\item{tolerancia_simplificacion}{Tolerancia de simplificacion en metros (def: 100 metros).}
}
\value{
Un dataframe estandarizado con los registros encontrados.
}
\description{
Busca ocurrencias de especies en GBIF dentro de un poligono de distrito
}
