% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_inat.R
\name{buscar_inat_por_poligono}
\alias{buscar_inat_por_poligono}
\title{Busca ocurrencias de especies en iNaturalist dentro de un poligono de distrito o provincia}
\usage{
buscar_inat_por_poligono(
  poligono_sf,
  query = NULL,
  taxon_name = NULL,
  grupo = NULL,
  calidad = "research",
  limite = 500
)
}
\arguments{
\item{poligono_sf}{Objeto sf que representa la unidad espacial (EPSG:4326).}

\item{query}{Texto libre de busqueda (opcional).}

\item{taxon_name}{Nombre de la especie o grupo taxonomico (opcional).}

\item{grupo}{Grupo taxonomico: "flora", "fauna" o NULL.}

\item{calidad}{Grado de calidad: "research" (por defecto), "casual" o NULL (ambos).}

\item{limite}{Numero maximo de registros a recuperar (max. 10000 para get_inat_obs).}
}
\value{
Un dataframe estandarizado con los registros encontrados que caen dentro del poligono.
}
\description{
Busca ocurrencias de especies en iNaturalist dentro de un poligono de distrito o provincia
}
