% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{graficar_ocurrencias}
\alias{graficar_ocurrencias}
\title{Grafica el mapa de ocurrencias sobre el poligono de la unidad administrativa}
\usage{
graficar_ocurrencias(
  resultado_lista,
  color_por = "source",
  guardar_mapa = FALSE
)
}
\arguments{
\item{resultado_lista}{Lista de resultados devuelta por buscar_especies_peru(), buscar_especies_distrito() o buscar_especies_provincia().}

\item{color_por}{Columna para clasificar los colores ("source" por defecto, o "kingdom").}

\item{guardar_mapa}{Logico; si es TRUE guarda el grafico en results/ como PNG (def: FALSE).}
}
\value{
El objeto de grafico ggplot.
}
\description{
Grafica el mapa de ocurrencias sobre el poligono de la unidad administrativa
}
