% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{verificar_y_configurar_entorno}
\alias{verificar_y_configurar_entorno}
\title{Verify package dependencies}
\usage{
verificar_y_configurar_entorno(instalar = FALSE)
}
\arguments{
\item{instalar}{Whether to install missing dependencies explicitly.}
}
\value{
\code{TRUE}, invisibly, when all dependencies are available.
}
\description{
Verify package dependencies
}
